# docs/swagger.yaml
definitions:
  db.GetTeamsByUserIDRow:
    properties:
      membership_id:
        type: integer
      team_domain:
        type: string
      team_id:
        type: integer
      team_name:
        type: string
      user_joined_at:
        type: string
      user_role:
        $ref: '#/definitions/db.TeamUserRole'
    type: object
  db.Tag:
    properties:
      key:
        type: string
      key_id:
        type: integer
      options:
        items:
          properties:
            id:
              type: integer
            value:
              type: string
          type: object
        type: array
      value:
        type: string
        x-nullable: true
      value_id:
        type: integer
        x-nullable: true
    type: object
  db.TagDataType:
    enum:
    - string
    - number
    - boolean
    type: string
    x-enum-varnames:
    - TagDataTypeString
    - TagDataTypeNumber
    - TagDataTypeBoolean
  db.TagKey:
    properties:
      created_at:
        type: string
      data_type:
        $ref: '#/definitions/db.TagDataType'
      id:
        type: integer
      name:
        type: string
      team_id:
        type: integer
    type: object
  db.TagValue:
    properties:
      created_at:
        type: string
      id:
        type: integer
      key_id:
        type: integer
      value:
        type: string
    type: object
  db.Team:
    properties:
      created_at:
        type: string
      domain:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  db.TeamUserRole:
    enum:
    - member
    - mod
    type: string
    x-enum-varnames:
    - TeamUserRoleMember
    - TeamUserRoleMod
  dto.CreateTagKeyBody:
    properties:
      data_type:
        allOf:
        - $ref: '#/definitions/db.TagDataType'
        maxLength: 100
        minLength: 2
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - data_type
    - name
    type: object
  dto.CreateTagKeyResponse:
    properties:
      data:
        $ref: '#/definitions/db.TagKey'
    type: object
  dto.CreateTagValueBody:
    properties:
      value:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - value
    type: object
  dto.CreateTagValueResponse:
    properties:
      data:
        $ref: '#/definitions/db.TagValue'
    type: object
  dto.DeleteTagKeyResponse:
    properties:
      data:
        $ref: '#/definitions/db.TagKey'
    type: object
  dto.DeleteTagValueResponse:
    properties:
      data:
        $ref: '#/definitions/db.TagValue'
    type: object
  dto.ErrorResponse:
    properties:
      message:
        example: Something went wrong
        type: string
    type: object
  dto.GetTeamsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/db.GetTeamsByUserIDRow'
        type: array
    type: object
  dto.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  dto.JoinTeamResponse:
    properties:
      data:
        $ref: '#/definitions/db.Team'
    type: object
  dto.ListFiltersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/db.Tag'
        type: array
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    type: object
  dto.LogoutResponse:
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.RegisterResponse:
    type: object
  dto.UpdateTagKeyBody:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
  dto.UpdateTagKeyResponse:
    properties:
      data:
        $ref: '#/definitions/db.TagKey'
    type: object
info:
  contact:
    email: dashskndash@gmail.com
    name: skndash96
  description: This is the backend server for Lastnight
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Lastnight API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      description: Authenticates a user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login endpoint
      tags:
      - Auth
  /api/auth/logout:
    delete:
      description: Logs out the current user by clearing the authentication token
        cookie
      produces:
      - application/json
      responses:
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Logout endpoint
      tags:
      - Auth
  /api/auth/register:
    post:
      description: Registers a new user and returns an authentication token
      parameters:
      - description: Register credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Signup endpoint
      tags:
      - Auth
  /api/health:
    get:
      description: Returns the health status of the application
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Healthcheck endpoint
      tags:
      - Health
  /api/teams:
    get:
      description: Get the user's teams list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetTeamsResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get Teams
      tags:
      - Team
  /api/teams/{teamID}/filters:
    get:
      description: Get the filter tags of a team
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListFiltersResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Filters
      tags:
      - Tag
  /api/teams/{teamID}/tags:
    post:
      description: Create a new tag key
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: Tag
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTagKeyBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateTagKeyResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: New Tag Key
      tags:
      - Tag
  /api/teams/{teamID}/tags/{tagID}:
    delete:
      description: Delete a tag key
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: tagID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteTagKeyResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete Tag Key
      tags:
      - Tag
    put:
      description: Update a tag key
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: tagID
        required: true
        type: string
      - description: Tag
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTagKeyBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateTagKeyResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update Tag Key
      tags:
      - Tag
  /api/teams/{teamID}/tags/{tagID}/values:
    post:
      description: Create a new tag value
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: tagID
        required: true
        type: string
      - description: Value
        in: body
        name: value
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTagValueBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateTagValueResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create Tag Value
      tags:
      - Tag
  /api/teams/{teamID}/tags/{tagID}/values/{valueID}:
    delete:
      description: Delete a tag value
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: Tag ID
        in: path
        name: tagID
        required: true
        type: string
      - description: Value ID
        in: path
        name: valueID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteTagValueResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete Tag Value
      tags:
      - Tag
  /api/teams/default:
    post:
      description: Join the default team for the user
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.JoinTeamResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Join Default Team
      tags:
      - Team
swagger: "2.0"