# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql/driver"
	"fmt"
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type TagDataType string

const (
	TagDataTypeString  TagDataType = "string"
	TagDataTypeNumber  TagDataType = "number"
	TagDataTypeBoolean TagDataType = "boolean"
)

func (e *TagDataType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TagDataType(s)
	case string:
		*e = TagDataType(s)
	default:
		return fmt.Errorf("unsupported scan type for TagDataType: %T", src)
	}
	return nil
}

type NullTagDataType struct {
	TagDataType TagDataType `json:"tag_data_type"`
	Valid       bool        `json:"valid"` // Valid is true if TagDataType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTagDataType) Scan(value interface{}) error {
	if value == nil {
		ns.TagDataType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TagDataType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTagDataType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TagDataType), nil
}

type TeamUserRole string

const (
	TeamUserRoleMember TeamUserRole = "member"
	TeamUserRoleMod    TeamUserRole = "mod"
)

func (e *TeamUserRole) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TeamUserRole(s)
	case string:
		*e = TeamUserRole(s)
	default:
		return fmt.Errorf("unsupported scan type for TeamUserRole: %T", src)
	}
	return nil
}

type NullTeamUserRole struct {
	TeamUserRole TeamUserRole `json:"team_user_role"`
	Valid        bool         `json:"valid"` // Valid is true if TeamUserRole is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTeamUserRole) Scan(value interface{}) error {
	if value == nil {
		ns.TeamUserRole, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TeamUserRole.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTeamUserRole) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TeamUserRole), nil
}

type Account struct {
	ID                int32     `json:"id"`
	UserID            int32     `json:"user_id"`
	Provider          string    `json:"provider"`
	ProviderAccountID string    `json:"provider_account_id"`
	Password          []byte    `json:"password"`
	CreatedAt         time.Time `json:"created_at"`
	UpdatedAt         time.Time `json:"updated_at"`
}

type MemberFilter struct {
	ID           int32     `json:"id"`
	MembershipID int32     `json:"membership_id"`
	KeyID        int32     `json:"key_id"`
	ValueID      int32     `json:"value_id"`
	CreatedAt    time.Time `json:"created_at"`
}

type Session struct {
	ID     pgtype.UUID `json:"id"`
	UserID int32       `json:"user_id"`
	Email  string      `json:"email"`
	Expiry time.Time   `json:"expiry"`
}

type TagKey struct {
	ID        int32       `json:"id"`
	TeamID    int32       `json:"team_id"`
	Name      string      `json:"name"`
	DataType  TagDataType `json:"data_type"`
	CreatedAt time.Time   `json:"created_at"`
}

type TagValue struct {
	ID        int32     `json:"id"`
	KeyID     int32     `json:"key_id"`
	Value     string    `json:"value"`
	CreatedAt time.Time `json:"created_at"`
}

type Team struct {
	ID        int32     `json:"id"`
	Name      string    `json:"name"`
	Domain    string    `json:"domain"`
	CreatedAt time.Time `json:"created_at"`
}

type TeamMembership struct {
	ID       int32        `json:"id"`
	TeamID   int32        `json:"team_id"`
	UserID   int32        `json:"user_id"`
	Role     TeamUserRole `json:"role"`
	JoinedAt time.Time    `json:"joined_at"`
}

type User struct {
	ID        int32     `json:"id"`
	Email     string    `json:"email"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
}